requirements.txt
flask==3.1.2
bootstrap-flask==2.5.0
gunicorn==23.0.0
pre-commit==4.5.1
markdown2==2.5.4
pydantic==2.12.5
python-frontmatter==1.1.0
